fw_ex/praw_spiked/mcp-reddit-server/pyproject.toml
[project]
name = "mcp-reddit-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "asyncpraw>=7.8.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "praw>=7.8.1",
    "python-dotenv>=1.1.0",
]
